+++ 7. Group Project/W6D5_Project/W6D5_Project/Spring Hibernate/src/com/oocl/mnlbc/controller/OrderProductsController.java
@@ -73,7 +73,7 @@ public class OrderProductsController {
     * @return List<Product>
     */
    @RequestMapping(value = "/getOrderProductsByOrder", method = RequestMethod.GET)
-   public List<Product> getOrderProductsByOrder(@RequestParam int id) {
+   public List<Product> getOrderProductsByOrder(@RequestParam long id) {
       if (id >= 0) {
          return this.orderprodSVC.getOrderProductsByOrder(id);
       }
