+++ 7. Group Project/W6D5_Project/W6D5_Project/Spring Hibernate/src/com/oocl/mnlbc/dao/inf/ProductDAO.java
@@ -14,6 +14,6 @@ public interface ProductDAO {
 
 	public List<Product> getProductByCategory(String cat);
 
-	public int removeProduct(int id);
+	public int removeProduct(long id);
 
 }
