+++ 7. Group Project/W6D5_Project/W6D5_Project/Spring Hibernate/src/com/oocl/mnlbc/svc/inf/ProductSVC.java
@@ -13,7 +13,7 @@ public interface ProductSVC {
 
 	public List<Product> getProductsByCategory(String category);
 
-	public int removeProduct(int id);
+	public int removeProduct(long id);
 
 	// public int updateStock(String[] array);
 	// public Product getProduct(String id);
